Add script to remove references to tcl / tk
Script post-processes matplotlib wheels to remove libtcl / libtk from
the NEEDED sections of _tkagg.so, and remove vendored tcl / tk libs.

This means that the TCL / Tk symbols come from the Python process at
runtime, via the Tkinter import.

This is a bit risky, because it depends on TCL and Tk having the same
ABI across versions for the symbols we use, but testing suggests that it
works OK when compiling against tcl / tk 8.4 and running with 8.6, so
it's a reasonable guess that we'll be OK for a while.